requirements.txt
# Data processing and analysis
pandas>=1.5.0
numpy>=1.20.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Model training and evaluation
scikit-learn>=1.0.0
joblib>=1.1.0

# Data file handling
openpyxl>=3.0.0

# Dashboard and visualization - explicitly pinned versions to avoid conflicts
streamlit>=1.10.0
plotly==5.14.1
tenacity>=8.1.0  # Required by plotly

# Additional dashboard dependencies
watchdog>=2.2.0  # For improved file watching in Streamlit

# Audio processing (optional, for future development)
# librosa>=0.9.0

# Cloud functionality - uncomment for GCP deployment
google-cloud-storage>=2.0.0
google-cloud-firestore>=2.7.0
google-cloud-functions>=1.8.0

# Alerting
twilio==8.0.0
sendgrid==6.10.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development
black==23.3.0
flake8==6.0.0
